requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.4.0
astroid==3.2.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
av==13.0.0
axial-positional-embedding==0.2.1
backcall==0.2.0
bitarray==2.9.2
braceexpand==0.1.7
certifi==2024.2.2
cffi==1.16.0
chamfer==2.0.0
charset-normalizer==3.3.2
click==8.1.7
clip==1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
debugpy==1.6.5
decorator==5.1.1
decord==0.6.0
dgl-cu117==0.9.1.post1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
dukpy==0.4.0
easydict==1.13
einops==0.8.0
emd-ext==0.0.0
entrypoints==0.4
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
flash-attn==2.6.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
hf-transfer==0.1.6
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.25.1
hydra-core==1.0.7
idna==3.7
imageio==2.35.1
importlib-metadata==7.1.0
importlib-resources==6.4.0
iopath==0.1.10
ipykernel==6.29.3
ipython==8.12.0
ipywidgets==8.1.2
isort==5.13.2
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
jupyter-client==7.3.4
jupyter-core==5.7.2
jupyterlab-widgets==3.0.10
kiwisolver==1.4.5
KNN-CUDA==0.2
lazy-loader==0.4
littleutils==0.2.2
llvmlite==0.42.0
lmdb==1.4.1
local-attention==1.9.1
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
multidict==6.0.5
multimethod==1.12
mutf8==1.0.6
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numba==0.59.1
numpy==1.26.4
ogb==1.3.6
omegaconf==2.0.6
open-clip-torch==2.26.1
openai==1.35.10
opencv-python==4.9.0.80
outdated==0.2.2
packaging==24.0
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
peft==0.11.1
performer-pytorch==1.1.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
Pillow-SIMD==9.5.0.post1
platformdirs==4.2.1
pointnet2-cuda==0.0.0
pointnet2-ops==3.0.0
portalocker==2.8.2
prompt-toolkit==3.0.42
protobuf==4.25.3
psutil==5.9.8
ptflops==0.7.2.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
pygments==2.18.0
pylint==3.2.7
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorchvideo==0.1.5
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rdkit-pypi==2022.9.5
regex==2024.5.10
requests==2.31.0
sacrebleu==2.4.2
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==3.1.0
sentry-sdk==2.1.1
setproctitle==1.3.3
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.2
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.11
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.13.2
torch==1.13.1+cu117
torch-geometric==2.5.3
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-tb-profiler==0.4.3
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tornado==6.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.18.2
wcwidth==0.2.13
webdataset==0.2.100
werkzeug==3.0.3
wget==3.2
widgetsnbextension==4.0.10
yacs==0.1.8
yarl==1.9.4
you-get==0.4.1730
zipp==3.18.1
